+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 3. Removed `api` directory in favor of `service` directory
 4. Upgrade discord.js v12 -> v13
 5. Add /help bounty, fix grammar for text, simplify bot commands
+6. Capture bounty description and criteria from bot interaction messages
 
 ## 1.2.1-RELEASE (2021-08-12)